+++ CHANGELOG.md
@@ -8,6 +8,8 @@ backwards compatibility.
 
 * Blink the signal LED twice to signalize new version.
 * Add harmonic minor mode.
+* Introduce configuration menu.
+* Enable overdrive through the configuration menu.
 * Extend the build manual with additional notes.
 * Update pictures in the build manual to match version 2.